feat: :sparkles: Added Pagination component and integrated it in HomePage
I have created the Pagination.jsx component to manage page navigation for the cryptocurrency data.
It includes handlers for navigating to the previous and next pages, with constraints to prevent going below page 1 or above page 10.
The current page is highlighted in red. Additionally, I imported the Pagination component into HomePage.jsx to allow users to navigate through different pages of cryptocurrency data.