feat: :sparkles: Made currency dynamic and added Search component
I have updated the cryptoApi.js file to make the currency parameter dynamic, allowing it to be set based on user input.
A new Search.jsx component was created to provide users with an input field and a dropdown menu to select different currencies (USD, EUR, JPY).
This selection updates the state in HomePage, enabling the display of cryptocurrency prices in the chosen currency.