feat: :sparkles: Added chart display functionality with dynamic toggle
I added a chart state in the HomePage.jsx component to handle the visibility of the chart.
The chart is toggled when a row in the TableCoin.jsx is clicked, using the setChart method passed through the components.
I also created a new component Chart.jsx to display the chart, with a close button to hide it when clicked.
The chart container is styled with a fixed position, backdrop blur effect, and a sleek design using Chart.module.css.